Improved rig and bone pickers. Initial offset calculator is now interactive
World Rig and World Bone now let you pick from the objects and bones possible, so no more typos if you make a mistake.

Relative Rig and Relative Bone has been renamed to Initial Offsets Rig and Bone respectively. This is to make it clearer that t hey are for calculating the initial offsets.

The are also updated dynamically now, too, and also have the same ui improvements as World Rig and World Bone.