New features in creating timestamps. Fix actor bone hash when not fully lowercase
Added "e" and "r" suffixes for timestamps.

e stands for evaluate, meaning it expects python code. Useful if you want to do math for the timestamp because you're combining/separating animations and don't want to pull out a calculator.

"r" stands for relative meaning that the current timestamp is already relative to the current clip, so no need for offsetting.