Blender specific arguments are no longer accepted in ClipResource
This is a change meant to make creating clips through code much easier.

Instead of accepting blender specific arguments in the __init__, it now requires the passed arguments to be the ones used for export.

The original logic has now been moved to NewClipExporter since that's the code for the blender scene.